Code for version 1 of the robot
Just like the STL files for version 1 of the robot, this code is utterly useless. It's a mess of code that was taking from the Adafruit website. The code was never finished because I finally noticed that the motors I was using for the knee joints did not supply enough torque. Nevertheless, here it is if anyone wants to look at it for some reason.